check_naf PNP templ: Print nav* totals at bottom of the legend.
In the case of navi* (INodes), the total was not previously included in the
legend at all, making it easy to mix up the 'total' line with the 'used' value
in case a very small percentage was in use.

In general, printing the total at the bottom imho makes the graph easier to
read, as the legend now describes the graph from bottom to top.