Added initial version of check_junos_bgp.pl.
The plugin connects to a Juniper Router and queries BGP routing information
using the 'show bgp neighbor' command. It then executes any of the following
checks:

 * peers_count: Total number of peers.
 * prefix_count: Number of active prefixes of single peers.

Each check may consider a subset of peers only. This is done by specifying a
"target" which may either be an IPv4/IPv6 address, a regex or a string. In the
latter two cases, the regex/string is compared with the peer description.

Warning and critical ranges may be specified for each single check.

The format for specifying checks is:
checkname[,target[,warning[,critical]]]