check_junos: Added 'interface_forwarding' check.
Check the forwarding state of interfaces as provided by 'show
ethernet-switching interfaces'. Storm control, MAC limit and BPDUs will be
considered CRITICAL states. If a target is specified, only the specified
interface(s) will be taken into account. Targets may be specified as
<interface_name>:<forwarding_state> in which case a CRITICAL state is assumed
if the specified interface is not in the specified state.